Add VolumeUtils for slice positioning, clean up examples accordingly
