+++ src/test/kotlin/graphics/scenery/tests/examples/cluster/DemoReelExample.kt
@@ -12,8 +12,6 @@ import graphics.scenery.attribute.material.Material
 import graphics.scenery.volumes.Colormap
 import graphics.scenery.volumes.TransferFunction
 import graphics.scenery.volumes.Volume
-import net.imglib2.type.numeric.integer.UnsignedByteType
-import net.imglib2.type.numeric.integer.UnsignedShortType
 import org.scijava.ui.behaviour.ClickBehaviour
 import java.nio.file.Paths
 import kotlin.concurrent.thread
@@ -71,8 +69,7 @@ class DemoReelExample: SceneryBase("Demo Reel") {
 
         val histoneVolume = Volume.fromPathRaw(
             Paths.get("$driveLetter:/ssd-backup-inauguration/CAVE_DATA/histones-isonet/stacks/default/"),
-            hub,
-            UnsignedShortType()
+            hub
         )
         histoneVolume.transferFunction = TransferFunction.ramp(0.1f, 1.0f)
         histoneVolume.colormap = Colormap.get("hot")
@@ -82,8 +79,7 @@ class DemoReelExample: SceneryBase("Demo Reel") {
 
         val drosophilaVolume = Volume.fromPathRaw(
             Paths.get("$driveLetter:/ssd-backup-inauguration/CAVE_DATA/droso-royer-autopilot-transposed/"),
-            hub,
-            UnsignedShortType()
+            hub
         )
         drosophilaVolume.spatial {
             rotation.rotateX(1.57f)
@@ -96,8 +92,7 @@ class DemoReelExample: SceneryBase("Demo Reel") {
 
         val retinaVolume = Volume.fromPathRaw(
             Paths.get("$driveLetter:/ssd-backup-inauguration/CAVE_DATA/retina_test2/"),
-            hub,
-            UnsignedShortType()
+            hub
         )
         retinaScene.addChild(retinaVolume)
         retinaScene.visible = false
