+++ src/test/kotlin/graphics/scenery/tests/examples/volumes/SplitVolumeExample.kt
@@ -4,6 +4,7 @@ import graphics.scenery.*
 import graphics.scenery.backends.Renderer
 import graphics.scenery.controls.TrackedStereoGlasses
 import graphics.scenery.attribute.material.Material
+import graphics.scenery.utils.extensions.positionVolumeSlices
 import graphics.scenery.volumes.Colormap
 import graphics.scenery.volumes.TransferFunction
 import graphics.scenery.volumes.Volume
@@ -49,8 +50,6 @@ class SplitVolumeExample: SceneryBase("Split volume data", 1280, 720) {
         val pair = Volume.fromPathRawSplit(Paths.get(getDemoFilesPath() + "/volumes/box-iso/box_200_200_200.raw"), hub = hub, type = UnsignedShortType(), sizeLimit = 20000000)
         val parent = pair.first as RichNode
         val volumeList = pair.second
-
-//        Volume.positionSlices(volumeList, volumeList.first().pixelToWorldRatio)
 
         parent.positionVolumeSlices(volumeList)
