+++ src/test/kotlin/graphics/scenery/tests/examples/volumes/VolumeExample.kt
@@ -7,8 +7,6 @@ import graphics.scenery.attribute.material.Material
 import graphics.scenery.volumes.Colormap
 import graphics.scenery.volumes.TransferFunction
 import graphics.scenery.volumes.Volume
-import net.imglib2.type.numeric.integer.UnsignedByteType
-import net.imglib2.type.numeric.integer.UnsignedShortType
 import org.joml.Vector3f
 import java.nio.file.Paths
 import kotlin.concurrent.thread
@@ -48,7 +46,7 @@ class VolumeExample: SceneryBase("Volume Rendering example", 1280, 720) {
         s.material().diffuse = Vector3f(0.0f, 0.0f, 0.0f)
         scene.addChild(s)
 
-        val volume = Volume.fromPathRaw(Paths.get(getDemoFilesPath() + "/volumes/box-iso/"), hub, UnsignedShortType())
+        val volume = Volume.fromPathRaw(Paths.get(getDemoFilesPath() + "/volumes/box-iso/"), hub)
         volume.name = "volume"
         volume.colormap = Colormap.get("viridis")
         volume.spatial {
